code refactoring and cleaning
