+++ sporniket-javabeans-doclet/src/main/java/com/sporniket/libre/javabeans/doclet/expander/JavaSourceGenerator.java
@@ -1,0 +1,33 @@
+/**
+ *
+ */
+package com.sporniket.libre.javabeans.doclet.expander;
+
+/**
+ * Interface of a java source code generator.
+ *
+ * @author dsporn
+ *
+ */
+public interface JavaSourceGenerator
+{
+	default void generate()
+	{
+		outputPackageStatement();
+		outputImportStatements();
+
+		outputClassBegin();
+		outputClassBody();
+		outputClassEnd();
+	}
+
+	void outputClassBegin();
+
+	void outputClassBody();
+
+	void outputClassEnd();
+
+	void outputImportStatements();
+
+	void outputPackageStatement();
+}
